Добавил меню работы со статьями

+++ src/main/java/org/oop/api/IArticleService.java
@@ -6,10 +6,10 @@ import java.util.List;
 
 public interface IArticleService {
     Article createArticle(String title, String content);
-    Article getArticleById(int id);
+    Article getArticleById(long id);
     List<Article> getArticlesByTitle(String title);
     List<Article> getAllArticles();
-    boolean updateArticle(int id, String title, String content);
-    boolean deleteArticle(int id);
+    boolean updateArticle(long id, String title, String content);
+    boolean deleteArticle(long id);
 }
