+++ src/main/java/org/oop/api/dao/IArticleDao.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface IArticleDao {
     Article createArticle(Article article);
-    Article getArticleById(int id);
+    Article getArticleById(long id);
     List<Article> getArticlesByTitle(String title);
     List<Article> getAllArticles();
     boolean updateArticle(Article article);
-    boolean deleteArticle(int id);
+    boolean deleteArticle(long id);
 }
